リファクタ modal化してchannelCreate modalの下に設置した
